Temporary fix for duplicate label generation
Sometimes the default goto is already handled by the conditional gotos and the generator would emit the same label twice. The current fix is to always rename the default case, but a better fix would be to not emit the default case in the first place if a conflict is present.